+++ fp-macros/src/lib.rs
@@ -21,8 +21,8 @@ pub(crate) mod generate;
 pub(crate) mod impl_kind;
 pub(crate) mod parse;
 
-#[cfg(test)]
-mod property_tests;
+// #[cfg(test)]
+// mod property_tests;
 
 /// Generates the name of a Kind trait based on its signature.
 ///